refactor: Renamed package and files from 'solver' to 'helper'
