+++ runelite-plugin.properties
@@ -1,5 +1,5 @@
-displayName=Random Event Solver
+displayName=Random Event Helper
 author=infinitay
 description=A plugin to help solve various random events
 tags=random,event,helper,solver,exam,surprise exam,quiz,bee,beehive,beekeeper,freaky forester,pheasant,pinball,flippa,drill demon,damien,exercise,gravedigger,coffin
-plugins=randomeventsolver.RandomEventSolverPlugin
+plugins=randomeventshelper.RandomEventHelperPlugin
