+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'randomeventsolver'
+rootProject.name = 'randomeventhelper'
