--- src/main/java/randomeventsolver/RandomEventSolverConfig.java
@@ -1,70 +1,0 @@
-package randomeventsolver;
-
-import net.runelite.client.config.Config;
-import net.runelite.client.config.ConfigGroup;
-import net.runelite.client.config.ConfigItem;
-import net.runelite.client.eventbus.Subscribe;
-
-@ConfigGroup("randomeventhelper")
-public interface RandomEventSolverConfig extends Config
-{
-	@ConfigItem(
-		keyName = "isSurpriseExamEnabled",
-		name = "Surprise Exam",
-		description = "Helps highlight the answers for the Surprise Exam random event. Supports both matching and next item questions."
-	)
-	default boolean isSurpriseExamEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "isBeekeeperEnabled",
-		name = "Beekeeper",
-		description = "Helps highlight the answers for the Beekeeper random event."
-	)
-	default boolean isBeekeeperEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "isFreakyForesterEnabled",
-		name = "Freaky Forester",
-		description = "Helps highlight the answers for the Freaky Forester random event."
-	)
-	default boolean isFreakyForesterEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "isPinballEnabled",
-		name = "Pinball",
-		description = "Helps highlight the answers for the Pinball random event."
-	)
-	default boolean isPinballEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "isDrillDemonEnabled",
-		name = "Drill Demon",
-		description = "Helps highlight the answers for the Drill Demon random event."
-	)
-	default boolean isDrillDemonEnabled()
-	{
-		return true;
-	}
-
-	@ConfigItem(
-		keyName = "isGravediggerEnabled",
-		name = "Gravedigger",
-		description = "Helps highlight the answers for the Gravedigger random event."
-	)
-	default boolean isGravediggerEnabled()
-	{
-		return true;
-	}
-}
