--- src/main/java/randomeventsolver/RandomEventSolverItemOverlay.java
@@ -1,29 +1,0 @@
-package randomeventsolver;
-
-import java.awt.Graphics2D;
-import javax.inject.Inject;
-import net.runelite.api.Client;
-import net.runelite.api.widgets.WidgetItem;
-import net.runelite.client.ui.overlay.WidgetItemOverlay;
-import net.runelite.client.ui.overlay.outline.ModelOutlineRenderer;
-
-public class RandomEventSolverItemOverlay extends WidgetItemOverlay
-{
-	private final Client client;
-	private final RandomEventSolverPlugin plugin;
-	private final ModelOutlineRenderer modelOutlineRenderer;
-
-	@Inject
-	public RandomEventSolverItemOverlay(Client client, RandomEventSolverPlugin plugin, ModelOutlineRenderer modelOutlineRenderer)
-	{
-		this.client = client;
-		this.plugin = plugin;
-		this.modelOutlineRenderer = modelOutlineRenderer;
-		showOnInventory();
-	}
-
-	@Override
-	public void renderItemOverlay(Graphics2D graphics, int itemId, WidgetItem widgetItem)
-	{
-	}
-}
