--- src/main/java/randomeventsolver/RandomEventSolverOverlay.java
@@ -1,38 +1,0 @@
-package randomeventsolver;
-
-import java.awt.Dimension;
-import java.awt.Graphics2D;
-import java.awt.image.BufferedImage;
-import javax.inject.Inject;
-import net.runelite.api.Client;
-import net.runelite.client.game.SpriteManager;
-import net.runelite.client.ui.overlay.Overlay;
-import net.runelite.client.ui.overlay.OverlayLayer;
-import net.runelite.client.ui.overlay.OverlayPosition;
-import net.runelite.client.ui.overlay.outline.ModelOutlineRenderer;
-
-public class RandomEventSolverOverlay extends Overlay
-{
-
-	private final Client client;
-	private final SpriteManager spriteManager;
-	private final RandomEventSolverPlugin plugin;
-	private final ModelOutlineRenderer modelOutlineRenderer;
-
-	@Inject
-	public RandomEventSolverOverlay(Client client, RandomEventSolverPlugin plugin, ModelOutlineRenderer modelOutlineRenderer, SpriteManager spriteManager)
-	{
-		this.client = client;
-		this.plugin = plugin;
-		this.modelOutlineRenderer = modelOutlineRenderer;
-		this.spriteManager = spriteManager;
-		setPosition(OverlayPosition.DYNAMIC);
-		setLayer(OverlayLayer.ABOVE_SCENE);
-	}
-
-	@Override
-	public Dimension render(Graphics2D graphics2D)
-	{
-		return null;
-	}
-}
