--- src/main/java/randomeventsolver/data/RandomEventItem.java
@@ -1,131 +1,0 @@
-package randomeventsolver.data;
-
-import com.google.common.collect.ImmutableMap;
-import java.util.Map;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum RandomEventItem
-{
-	AIR_RUNE(41168),
-	SQUARE_SHIELD_1(41188),
-	APRON(41190),
-	ARROWS(41177),
-	BATTLE_AXE(41176),
-	BANANA(41222),
-	TUNA(41209),
-	BEER(41152),
-	GRAPES(41207),
-	RUNE_OR_ESSENCE(41182),
-	BOOK(41181),
-	BREAD(41172),
-	CAKE(41202),
-	CANDLE_ON_STAND(27102),
-	CHEESE(41161),
-	CHEFS_HAT(41203),
-	COCKTAIL_SHAKER(27096),
-	TROUT_COD_PIKE_SALMON_1(41204),
-	CROSSBOW(41146),
-	CUP_OF_TEA(41162),
-	BOTTLE(41175),
-	EARTH_RUNE(41157),
-	EYE_PATCH(41165),
-	FEATHER(41224),
-	FIRE_RUNE(41215),
-	FROG_MASK(27101),
-	COCKTAIL_1(27097),
-	FULL_HELM(41178),
-	ONION(41226),
-	GEM_WITH_CROSS(41151),
-	GIN_OR_RUM(41219),
-	HAMMER(41183),
-	HARPOON(41158),
-	AXE(41184),
-	HERRING_OR_MACKEREL(41193),
-	HIGHWAYMAN_MASK(41195),
-	JESTER_HAT(41196),
-	JUG(41225),
-	KITESHIELD(41200),
-	CANDLE_LANTERN(41229),
-	LEDERHOSEN_HAT(41164),
-	CAPE_OF_LEGENDS(41167),
-	LOGS(41232),
-	LONGBOW(41198),
-	LONGSWORD(41150),
-	MACE(41211),
-	MED_HELM(41189),
-	BAR(41153),
-	MIME_MASK(41191),
-	NECKLACE(41216),
-	NEEDLE(41199),
-	ORE(41170),
-	PICKAXE(41194),
-	PIE(41205),
-	TROUT_COD_PIKE_SALMON_2(41206),
-	PINEAPPLE(41214),
-	PIRATE_BOOTS(41186),
-	PIRATE_HAT(41187),
-	PIRATE_HOOK(41228),
-	PIZZA(41185),
-	PLANT_POT(41208),
-	PLATELEGS(41179),
-	POT(41223),
-	RAKE(41212),
-	LEATHER_BOOTS(41220),
-	RING(27091),
-	WOODEN_SHIELD(41221),
-	PLATEBODY(27094),
-	TROUT_COD_PIKE_SALMON_3(41163),
-	HOLY_SYMBOL(41159),
-	FISH(41173),
-	SCIMITAR(41192),
-	SECATEURS(41197),
-	COCKTAIL_2(28421),
-	SHARK(41166),
-	SHEARS(41227),
-	SHORT_BOW(41171),
-	SHRIMP(41147),
-	BOOTS(27104),
-	FIGHTER_BOOTS(41160),
-	SPADE(41155),
-	SQUARE_SHIELD_2(41169),
-	STAFF(41174),
-	STRAWBERRY(41230),
-	KITCHEN_KNIFE(41201),
-	THREAD(41218),
-	TIARA(41148),
-	TINDERBOX(41154),
-	TROUT_COD_PIKE_SALMON_4(41217),
-	GARDENING_TROWEL(41210),
-	BASS(41180),
-	POTION(41149),
-	WATERING_CAN(41213),
-	WATERMELON_SLICE(41156),
-	WATER_RUNE(41231),
-	BONES(2674),
-
-	KEY(29232);
-
-	private final int modelID;
-
-	public static RandomEventItem fromModelID(int modelID)
-	{
-		return ITEM_MODEL_ID_MAP.get(modelID);
-	}
-
-	private static final Map<Integer, RandomEventItem> ITEM_MODEL_ID_MAP;
-
-	static
-	{
-		ImmutableMap.Builder<Integer, RandomEventItem> itemModelIDBuilder = new ImmutableMap.Builder<>();
-
-		for (RandomEventItem randomEventItem : values())
-		{
-			itemModelIDBuilder.put(randomEventItem.getModelID(), randomEventItem);
-		}
-
-		ITEM_MODEL_ID_MAP = itemModelIDBuilder.build();
-	}
-}
