+++ src/test/java/randomeventhelper/RandomEventHelperPluginTest.java
@@ -1,0 +1,13 @@
+package randomeventhelper;
+
+import net.runelite.client.RuneLite;
+import net.runelite.client.externalplugins.ExternalPluginManager;
+
+public class RandomEventHelperPluginTest
+{
+	public static void main(String[] args) throws Exception
+	{
+		ExternalPluginManager.loadBuiltin(RandomEventHelperPlugin.class);
+		RuneLite.main(args);
+	}
+}
