--- src/test/java/randomeventsolver/RandomEventSolverPluginTest.java
@@ -1,13 +1,0 @@
-package randomeventsolver;
-
-import net.runelite.client.RuneLite;
-import net.runelite.client.externalplugins.ExternalPluginManager;
-
-public class RandomEventSolverPluginTest
-{
-	public static void main(String[] args) throws Exception
-	{
-		ExternalPluginManager.loadBuiltin(RandomEventSolverPlugin.class);
-		RuneLite.main(args);
-	}
-}
